Claude refactor of UIManager
